Modular structure, added ViT_B_16 and ViT_B_32
